add fixed_point_finder and recurrentgemma
adde the FixedPointFinder repo,and the RecurrentGemma repo.
The code is already slightly modified from the original versions, specifically including the griffin_wrappers that wrapps griffin modules for use with the fixed_point_finder.

At this point using the GriffinRecurrentBlock should run without exceptions and train, but follows the input directly and not the desired "memory" output.