adjustments for ModifiedRecurrentBlock
added the ModifiedRecurrentBlock, which is a variation of the original RecurrentBlock that also returns the state trajectory of the RGLRU unit, together with the output and the cache.
Modified the other classes and scripts to use this.
run_FlipFlop should now train with this module and also plot the results.